Added authoriation conditions to my blog endpoints. Like, only the author can update and delete blogposts, and only the author can view his unpublished blogposts

+++ api/view/blogpost.py
@@ -11,12 +11,14 @@ from rest_framework import permissions
 
 class BlogList(APIView): 
     """ api/blogposts/
-        - GET list of
-            if authenticated: list of all blogposts by the user (even unpublished)
+        - GET list of (AUTH optional)
+            if authenticated: list of all blogposts by the user (even unpublished ones)
             else: list of all published blogposts
-        - POST a new blogpost (REQUIRES AUTH)
+        - POST a new blogpost (AUTH required)
+            the request.user will become the author
     """
     #permission_classes = [permissions.IsAuthenticated]
+
     def get(self, request, format=None):
         if request.user.is_authenticated:
             # if authenticated, will return all the posts of the user
@@ -28,45 +30,62 @@ class BlogList(APIView):
             blogposts = BlogPost.objects.all().filter(published_date__lte=timezone.now())
             serializer = BlogpostSerializer(blogposts, many=True)
             return Response(serializer.data)
-        
-
+    
     def post(self, request, format=None):
-        if request.user.is_authenticated:
+        if request.user.is_authenticated: # AUTH REQUIRED
             serializer = BlogpostSerializer(data=request.data)
             if serializer.is_valid():
+                serializer.validated_data["author"] = request.user
                 serializer.save()
                 return Response(serializer.data, status=status.HTTP_201_CREATED)
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        else:
+        else: #otherwise return error
             return Response({"detail":"Authentication required"}, status.HTTP_401_UNAUTHORIZED)
 
-
-    
 
 class BlogDetail(APIView):
     """api/blogpost/<int>/
        - GET a single blogpost
-       - PUT update a single blogpost
-       - DELETE a single blogpost
+            anyone can access any published post
+            but only the author can access their unpublished posts
+       - PUT update a single blogpost (AUTH required)
+            requires authentication and that the user should be the author of that post
+            the request.user will be the author of that post
+       - DELETE a single blogpost (AUTH required)
+            requires authentication and that the user should be the author of that post
     """
-    def get_object(self,pk):
+    def get_object(self,pk)->BlogPost:
         try:
             return BlogPost.objects.get(pk=pk)
         except BlogPost.DoesNotExist:
             raise Http404
     
     def get(self, request, pk, format=None):
-        serializer = BlogpostSerializer(self.get_object(pk))
+        blogpost = self.get_object(pk)
+        serializer = BlogpostSerializer(blogpost)
+        if (blogpost.published_date is None) and (request.user!=blogpost.author): #post is unpublished
+            return Response(data={"detail":"You are not authorized to view this unpublished post"},status=status.HTTP_403_FORBIDDEN)
         return Response(serializer.data)
 
     def put(self, request, pk, format=None):
-        serializer = BlogpostSerializer(self.get_object(pk), data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        owner_of_postdelete = self.get_object(pk).author
+        if request.user==owner_of_postdelete:
+            #authorized user is the author of the post
+            #the author will be the user
+            serializer = BlogpostSerializer(self.get_object(pk), data=request.data)
+            if serializer.is_valid():
+                serializer.validated_data["author"] = request.user
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response(data={"detail":"You are not authorized to PUT this blogpost"}, status=status.HTTP_403_FORBIDDEN)
     
     def delete(self, request, pk, format=None):
-        self.get_object(pk).delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
+        owner_of_postdelete = self.get_object(pk).author
+        if request.user==owner_of_postdelete:
+            #authorized user is the author of the post
+            self.get_object(pk).delete()
+            return Response(data={"detail":f"Deleted post {pk} successfully"}, status=status.HTTP_200_OK)
+        else:
+            return Response(data={"detail":"You are not authorized to DELETE this blogpost"}, status=status.HTTP_403_FORBIDDEN)
